chore: enable CPD detection (pmd.cpd.min 100000 -> 100), de-dup and baseline
pmd.cpd.min=100000 effectively disabled PMD's copy/paste detector: cpd-check
passed on virtually any duplication. Lower the threshold to 100 (PMD's
default) so CPD detects duplication going forward.

At 100 the reactor surfaces pre-existing duplications; each was assessed
for whether it is genuinely extractable logic or incidental/deliberate
similarity:

- Refactored: the identical isExtensionUpdateRequired guard chain in the
  Check validator/quickfix extension helpers lifted to
  AbstractCheckExtensionHelper (keyed on getExtensionPointId() plus a new
  protected getTargetClassName() hook); the identical private
  featureIterable helper in Abstract{,Streaming}FingerprintComputer
  extracted to a shared package-private FingerprintFeatures utility.
- Baselined with per-site CPD-OFF markers and reasons: test scaffolding in
  AbstractValidationTest, sibling test cases in ParameterListMatcherTest,
  incidental token overlap in DispatchingCheckImpl, migrated Xtend
  dispatch/emission code in FormatJvmModelInferrer (file-scoped) and
  FormatFragment2, and sibling test-class scaffolding in the two CheckCfg
  test classes.

The lifted guard chain keeps the overrides' public visibility: the hook is
deliberately public so tests can exercise it directly on injected helpers.

Full-reactor pmd:cpd-check is clean at threshold 100.

Closes #1339.
