+++ com.avaloq.tools.ddk.xtext/src/com/avaloq/tools/ddk/xtext/resource/AbstractFingerprintComputer.java
@@ -14,19 +14,15 @@ import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.Collections;
-import java.util.Iterator;
 import java.util.List;
 
-import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.common.util.URI;
 import org.eclipse.emf.ecore.EAttribute;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.EReference;
-import org.eclipse.emf.ecore.EStructuralFeature;
 import org.eclipse.emf.ecore.InternalEObject;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.util.EcoreUtil;
-import org.eclipse.emf.ecore.util.InternalEList;
 import org.eclipse.xtext.linking.lazy.LazyLinkingResource;
 
 import com.google.common.collect.Iterables;
@@ -210,43 +206,6 @@ public abstract class AbstractFingerprintComputer implements IFingerprintCompute
   }
 
   /**
-   * Return an Iterable containing all the contents of the given feature of the given object. If the
-   * feature is not many-valued, the resulting iterable will have one element. If the feature is an EReference,
-   * the iterable may contain proxies. The iterable may contain null values.
-   *
-   * @param <T>
-   *          The Generic type of the objects in the iterable
-   * @param obj
-   *          The object
-   * @param feature
-   *          The feature
-   * @return An iterable over all the contents of the feature of the object.
-   */
-  @SuppressWarnings("unchecked")
-  private <T> Iterable<T> featureIterable(final EObject obj, final EStructuralFeature feature) {
-    if (feature == null) {
-      return Collections.emptyList();
-    }
-    if (feature.isMany()) {
-      if (feature instanceof EAttribute || ((EReference) feature).isContainment()) {
-        return (Iterable<T>) obj.eGet(feature);
-      }
-      return new Iterable<T>() {
-        @Override
-        public Iterator<T> iterator() {
-          EList<T> list = (EList<T>) obj.eGet(feature);
-          if (list instanceof InternalEList<T> internalList) {
-            return internalList.basicIterator(); // Don't resolve
-          } else {
-            return list.iterator();
-          }
-        }
-      };
-    }
-    return Collections.singletonList((T) obj.eGet(feature, false)); // Don't resolve
-  }
-
-  /**
    * Generate a fingerprint for the target object using its URI.
    *
    * @param target
@@ -319,7 +278,7 @@ public abstract class AbstractFingerprintComputer implements IFingerprintCompute
     if (obj == null) {
       return NULL_STRING;
     }
-    final Iterable<? extends EObject> targets = featureIterable(obj, ref);
+    final Iterable<? extends EObject> targets = FingerprintFeatures.featureIterable(obj, ref);
     if (targets == null) {
       return NULL_STRING;
     }
@@ -370,7 +329,7 @@ public abstract class AbstractFingerprintComputer implements IFingerprintCompute
     if (obj == null) {
       return NULL_STRING;
     }
-    final Iterable<? extends Object> values = featureIterable(obj, attr);
+    final Iterable<? extends Object> values = FingerprintFeatures.featureIterable(obj, attr);
     if (values == null) {
       return NULL_STRING;
     }
@@ -468,7 +427,7 @@ public abstract class AbstractFingerprintComputer implements IFingerprintCompute
     if (obj == null) {
       return NO_EXPORT;
     }
-    final Iterable<? extends EObject> targets = featureIterable(obj, ref);
+    final Iterable<? extends EObject> targets = FingerprintFeatures.featureIterable(obj, ref);
     if (targets == null) {
       return NO_EXPORT;
     }
