+++ ddk-parent/pom.xml
@@ -32,7 +32,7 @@
     <!-- analysis parameter -->
     <compiler.showWarnings>true</compiler.showWarnings>
     <pmd.ruleset>${workspace}/ddk-configuration/pmd/ruleset.xml</pmd.ruleset>
-    <pmd.cpd.min>100000</pmd.cpd.min>
+    <pmd.cpd.min>100</pmd.cpd.min>
     <checkstyle.configLocation>${workspace}/ddk-configuration/checkstyle/avaloq.xml</checkstyle.configLocation>
     <checkstyle.test.configLocation>${workspace}/ddk-configuration/checkstyle/avaloq-test.xml</checkstyle.test.configLocation>
     <spotbugs.excludeFilterFile>${workspace}/ddk-configuration/findbugs/exclusion-filter.xml</spotbugs.excludeFilterFile>
